+++ .github/CHANGELOG.md
@@ -50,6 +50,12 @@
   ket = Gaussian(2).ket(max_prob=0.99, max_photons=3)
   ```
 
+* Gaussian transformations support a `bargmann` method for returning the bargmann representation. 
+  [(#239)](https://github.com/XanaduAI/MrMustard/pull/239)
+
+* BSGate.U now supports method='vanilla' (default) and 'schwinger' (slower, but stable to any cutoff)
+  [(#248)](https://github.com/XanaduAI/MrMustard/pull/248)
+
 ### Breaking Changes
 
 * The previous `callback` argument to `Optimizer.minimize` is now `callbacks` since we can now pass
@@ -72,17 +78,31 @@
 * Tests of the compact_fock module now use hypothesis.
   [(#235)](https://github.com/XanaduAI/MrMustard/pull/235)
 
+* Faster implementation of the fock representation of `BSgate`, `Sgate` and `SqueezedVacuum`, ranging from 5x to 50x.
+  [(#239)](https://github.com/XanaduAI/MrMustard/pull/239)
+
+* More robust implementation of cutoffs for States.
+[(#239)](https://github.com/XanaduAI/MrMustard/pull/239)
+
 ### Bug fixes
 
 * Fixed a bug that would make two progress bars appear during an optimization
   [(#235)](https://github.com/XanaduAI/MrMustard/pull/235)
 
-### Bug fixes
+* The displacement of the dual of an operation had the wrong sign
+  [(#239)](https://github.com/XanaduAI/MrMustard/pull/239)
+
+* When projecting a Gaussian state onto a Fock state, the upper limit of the autocutoff now respect the Fock projection.
+  [(#246)](https://github.com/XanaduAI/MrMustard/pull/246)
+
+* Fixed a bug for the algorithms that allow faster PNR sampling from Gaussian circuits using density matrices. When the 
+cutoff of the first detector is equal to 1, the resulting density matrix is now correct.
 
 ### Documentation
 
 ### Contributors
-[Filippo Miatto](https://github.com/ziofil), [Zeyue Niu](https://github.com/zeyueN)
+[Filippo Miatto](https://github.com/ziofil), [Zeyue Niu](https://github.com/zeyueN), 
+[Robbe De Prins](https://github.com/rdprins)
 
 ---